Rename package from L1Vertex to L1TVertex
